Combine networks into network files
