+++ ironfish/src/index.ts
@@ -6,9 +6,7 @@ export * from './assert'
 export * from './assets'
 export * from './blockchain'
 export * from './consensus'
-export * from './defaultNetworkDefinitions'
-export * from './networkDefinition'
-export { DEV_GENESIS_ACCOUNT } from './genesisBlocks/devnet'
+export * from './networks'
 export * from './chainProcessor'
 export * from './event'
 export * from './fileStores'
