+++ ironfish/src/node.ts
@@ -23,7 +23,7 @@ import { MiningManager } from './mining'
 import { PeerNetwork, PrivateIdentity, privateIdentityToIdentity } from './network'
 import { isHexSecretKey } from './network/identity'
 import { IsomorphicWebSocketConstructor, NodeDataChannelType } from './network/types'
-import { getNetworkDefinition } from './networkDefinition'
+import { getNetworkDefinition } from './networks'
 import { Package } from './package'
 import { Platform } from './platform'
 import { ALL_API_NAMESPACES, RpcMemoryClient } from './rpc'
